requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Statistical Analysis
statsmodels>=0.14.0

# Data Download
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Broker APIs (optional, install as needed)
ib-insync>=0.9.86
alpaca-trade-api>=3.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0
python-dotenv>=1.0.0

# Logging and Monitoring
colorlog>=6.7.0
